Reorder child terms

Reorder child terms

dag_reorder(dag, max_level = 2)

Arguments

dag

An ontology_Dag object.

max_level

Maximal depth of terms in DAG to apply reordering. The value should be set to a small integer because normally the effect of reordering too deep in the DAG is not noticable.

Value

An ontology_Dag object.

Details

For a given term, its child terms are reordered based on the numbers of cross-links of their offspring terms.

Denote c1 and c2 are two child terms, S1 and S2 are two sets of offspring terms of c1 and c2. We calculate k as the number of links between S1 and S2, which are the links linking parents in S1 and children in S2, or parents in S2 and children in S1. Denote n as the total number of links in the DAG, n - k is the distance bewteen c1 and c2 regarding how close the offspring sub-trees are connected.

With the distances from the set of child terms, TSP::TSP() is applied to reorder child terms.

Examples

# \donttest{
dag = create_ontology_DAG_from_GO_db()
dag_reorder(dag)
#> adjust child terms on level 0, 1 terms
#> adjust child terms on level 1, 21 terms
#> An ontology_DAG object:
#>   Source: GO BP / GO.db package 
#>   28140 terms / 56449 relations
#>   Root: GO:0008150 
#>   Terms: GO:0000001, GO:0000002, GO:0000003, GO:0000011, ...
#>   Max depth: 18 
#>   Avg number of parents: 2.01
#>   Aspect ratio: 342.43:1 (based on the longest distance to root)
#>                 780.22:1 (based on the shortest distance to root)
#>   Relations: is_a, part_of
#> 
#> With the following columns in the metadata data frame:
#>   id, name, definition
# }
1
#> [1] 1