Global options

Global options

simona_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)

Arguments

...

Name-value pairs for options.

RESET

Reset to default option values.

READ.ONLY

Only return read only options.

LOCAL

Only return local options.

ADD

Add new options.

Value

A single option value.

Details

There are the following global options:

use_cache: By default, information content of all terms is cached and reused. If use_cache is set to FALSE, IC will be re-calculated.

verbose: Whether to print messages?

anno_uniquify: In the annotation-based IC method, the union of items annotated to the term as well as all its offspring terms is used, which means the set of annotated items for the term is uniquified. If anno_uniquify is set to FALSE, the uniquification is not applied, we simply add the number of items annotated to the term and the numbers of items annotated to each of its offspring terms.

robot_jar: Path of the robot.jar file. The file can be found from https://github.com/ontodev/robot/releases.

To set an option, you can use $:

simona_opt$verbose = FALSE

or use it as a function:

simona_opt(verbose = FALSE)

Examples

simona_opt
#>  Option        Value
#>  -------------:-------
#>  use_cache     TRUE 
#>  verbose       TRUE 
#>  anno_uniquify TRUE 
#>  verbose       TRUE 
#>  robot_jar     NULL