Create the ontology_DAG object from the GO.db package

Create the ontology_DAG object from the GO.db package

create_ontology_DAG_from_GO_db(
  namespace = "BP",
  relations = "part of",
  org_db = NULL,
  evidence_code = NULL,
  verbose = simona_opt$verbose
)

Arguments

namespace

One of "BP", "CC" and "MF".

relations

Types of the GO term relations. In the GO.db package, the GO term relations can be "is_a", "part_of", "regulates", "negatively regulates", "positively regulates". Note since "regulates" is a parent relation of "negatively regulates", "positively regulates", if "regulates" is selected, "negatively regulates" and "positively regulates" are also selected. Note "is_a" is always included.

org_db

The name of the organism package or the corresponding database object, e.g. "org.Hs.eg.db" or directly the org.Hs.eg.db::org.Hs.eg.db object for human, then the gene annotation to GO terms will be added to the object. For other non-model organisms, consider to use the AnnotationHub package to find one.

evidence_code

A vector of evidence codes for gene annotation to GO terms. See https://geneontology.org/docs/guide-go-evidence-codes/.

verbose

Whether to print messages.

Value

An ontology_DAG object.

Examples

dag = create_ontology_DAG_from_GO_db()
#> 
#> relations: is_a, part_of
dag
#> An ontology_DAG object:
#>   Source: GO BP / GO.db package 3.17.0 
#>   27942 terms / 55956 relations
#>   Root: GO:0008150 
#>   Terms: GO:0000001, GO:0000002, GO:0000003, GO:0000011, ...
#>   Max depth: 18 
#>   Avg number of parents: 2.00
#>   Avg number of children: 1.88
#>   Aspect ratio: 363.92:1 (based on the longest distance from root)
#>                 782.78:1 (based on the shortest distance from root)
#>   Relations: is_a, part_of
#> 
#> With the following columns in the metadata data frame:
#>   id, name, definition