Generate a random DAG

Generate a random DAG

dag_random(
  n_children = c(2, 5),
  max = 1000,
  p = 0.1,
  power = 0.2,
  tree = FALSE,
  depth = NULL
)

Arguments

n_children

Range of the numbers of children on a term to sample.

max

Maximal number of terms.

p

Probability to connect to other terms.

power

The power.

tree

Whether the dag is a tree.

depth

Depth of the tree. Note tree and depth should be both set. dag(n_children = nc, tree = TRUE, depth = d) is a shortcut of dag(n_children = c(nc, nc), max = nc^(d+1)/(nc-1), p = 0). The setting of dag(n_children = 2, tree = TRUE, depth = d) generates a binary tree with max depth d and 2^d leaf nodes all have the same depth d.

Value

An ontology_DAG object.

Details

It first generates a tree growing from the root. On each node, it has a random number of children sampled from n_children and each child node can extend with a probability of 1/(height^power).

In the second step, each node on the tree has a probability of p to link to other terms in the tree.

The growing stops when the number of terms reaches max or there is no term can further extend.

Examples

dag_random()
#> An ontology_DAG object:
#>   Source: Ontology 
#>   999 terms / 1052 relations
#>   Root: 1 
#>   Terms: 1, 10, 100, 101, ...
#>   Max depth: 6 
#>   Avg number of parents: 1.05
#>   Aspect ratio: 95.33:1 (based on the longest distance to root)
#>                 90.67:1 (based on the shortest distance to root)
dag_random(p = 0)  # a tree
#> An ontology_DAG object:
#>   Source: Ontology 
#>   999 terms / 998 relations / a tree
#>   Root: 1 
#>   Terms: 1, 10, 100, 101, ...
#>   Max depth: 7 
#>   Aspect ratio: 84.86:1