Partition the DAG

Partition the DAG

partition_by_level(dag, level = 0, from = NULL, term_pos = NULL)

partition_by_k(dag, k)

Arguments

dag

An ontology_DAG object.

level

Depth in the DAG to cut. The DAG is cut below terms (or cut the links to their child terms) with depth == level.

from

A list of terms to cut. If it is set, level is ignored.

term_pos

Internally used.

k

Number of terms in a cluster. The splitting stops on a term if all its child-tree are smaller than k.

Details

Let's call the terms below the from term as "top terms" because they will be on top of the sub-DAGs after the partitioning. It is possible that a term in the middle of the DAG can be traced back to more than one top terms. To partition all terms exclusively, a term partitioned to the sub-DAG from the top term with the largest distance to the term. If a term has the same largest distances to several top terms, a random top term is selected.

In partition_by_k(), the DAG is first reduced to a tree where a child term only has one parent. The partition is done recursively by cutting into its child-trees. The splitting stops when all the child-trees have size less than k.

NA is assigned to the from terms, their ancestor terms, and terms having infinite directed distance to from terms.

Examples

if (FALSE) {
dag = create_ontology_DAG_from_GO_db()
pa = partition_by_level(dag)
table(pa)
pa = partition_by_k(dag, k = 1000)
table(pa)
}